requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
alembic
python-dotenv
passlib[bcrypt]
PyJWT
python-multipart
pydantic
email-validator
httpx
pytest
pytest-asyncio
pytest-cov
loguru
orjson